+++ CHANGELOG.md
@@ -10,6 +10,8 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 
 ### Added
 
+- Added get and delete functionality for contract definitions eclipse-tractusx/traceability-foss#1190
+
 ### Fixed
 
 - Access and Usage Policy Validation flow correction. #757
@@ -19,6 +21,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ### Changed
 
 - Replaced technical error message when trying to delete the configured default policy with a user-friendly message.
+- Renamed EdcCreateContractDefinitionRequest to EdcContractDefinition because it's used for both get and post eclipse-tractusx/traceability-foss#1190
 
 
 ## [5.2.0] - 2024-07-03